Locally, this drastically improved the performance of the Manage Packages page.

First load:
Before - 31 seconds
After - 8 seconds

Future loads:
Before - 21 seconds
After - 3 seconds