@joelverhagen 2018-12-12
Let's get a graph of page load durations for 10-20 iterations before and after.

Let's run against a real DB, like DEV DB.