@chenriksson 2018-12-12
Agree with @joelverhagen. I'd like to see average/percentile of before and after results, preferably quantified with the %diff (after - before / before * 100) calculation.

I also like to document the investigation by sharing a screenshot of the profile showing the inclusive %CPU before optimization. Of course, you also need to define the scenario that's being tested (e.g., lots of packages, lots of packages with lots of versions, etc).

For instance: https://github.com/NuGet/Engineering/issues/1453