@scottbommarito 2018-12-12
Unfortunately, it's not very easy for me to test an account that loads the page before the change locally against a SQL DB (due to AAD login), so I plan to gather data after this is deployed to DEV.