@chenriksson 2018-12-13
@scottbommarito You can test with AAD locally. In OneNote "Microsoft Logins (MSA) and NuGet Gallery", search for "Test Accounts for local usage". It would probably be nice if we stored this in the DEV AKV instead.